+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -1,9 +1,9 @@
 object Dependencies {
     object Versions {
-        const val kotlin = "1.8.10"
-        const val serialization = "1.4.0"
-        const val datetime = "0.4.0"
-        const val ktlintGradle = "10.2.0"
-        const val dokkaGradle = "1.5.0"
+        const val kotlin = "2.0.20"
+        const val serialization = "1.7.3"
+        const val datetime = "0.6.1"
+        const val ktlintGradle = "12.1.1"
+        const val dokkaGradle = "1.9.20"
     }
 }